refactor(#93, #94) : booking 메서드 락범위와 트랜잭션 범위 일치하도록 리팩토링, OrderCommandServiceTest, TicketCommandServiceTest, ShowDateCommandServiceTest, BookingServiceTest 단위테스트 작성 완료

+++ src/main/java/com/example/picket/domain/booking/service/BookingService.java
@@ -1,12 +1,9 @@
 package com.example.picket.domain.booking.service;
 
 import com.example.picket.common.enums.OrderStatus;
-import com.example.picket.common.enums.SeatStatus;
-import com.example.picket.common.enums.TicketStatus;
 import com.example.picket.common.exception.CustomException;
 import com.example.picket.domain.order.entity.Order;
 import com.example.picket.domain.order.service.OrderCommandService;
-import com.example.picket.domain.seat.entity.Seat;
 import com.example.picket.domain.seat.service.SeatQueryService;
 import com.example.picket.domain.seat_holding.service.SeatHoldingService;
 import com.example.picket.domain.show.entity.Show;
@@ -19,7 +16,6 @@ import com.example.picket.domain.ticket.service.TicketCommandService;
 import com.example.picket.domain.ticket.service.TicketQueryService;
 import com.example.picket.domain.user.entity.User;
 import com.example.picket.domain.user.service.UserQueryService;
-import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
@@ -32,83 +28,85 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
-import static org.springframework.http.HttpStatus.FORBIDDEN;
 
 @Service
 @RequiredArgsConstructor
 public class BookingService {
 
     private final RedissonClient redissonClient;
-    private final String KEY_PREFIX = "UPDATE-COUNT-LOCK:SHOW-DATE:";
+    private final String KEY_PREFIX = "BOOKING-LOCK:SHOW-DATE:";
 
     private final SeatHoldingService seatHoldingService;
 
-    private final SeatQueryService seatQueryService;
     private final ShowQueryService showQueryService;
     private final UserQueryService userQueryService;
     private final ShowDateQueryService showDateQueryService;
+    private final TicketQueryService ticketQueryService;
 
     private final OrderCommandService orderCommandService;
     private final TicketCommandService ticketCommandService;
-    private final TicketQueryService ticketQueryService;
     private final ShowDateCommandService showDateCommandService;
 
     @Transactional
     public Order booking(Long showId, Long showDateId, Long userId, List<Long> seatIds) throws InterruptedException {
 
-        Show foundShow = showQueryService.getShow(showId);
-        checkBookingTime(foundShow);
-
-        User foundUser = userQueryService.getUser(userId);
-
-        seatHoldingService.seatHoldingCheck(userId, seatIds);
-
-        ticketQueryService.checkTicketLimit(foundUser, foundShow);
-
-        List<Ticket> tickets = ticketCommandService.createTicket(foundUser, foundShow, seatIds);
-
-        Order order = orderCommandService.createOrder(foundUser, tickets);
-
-        // TODO : 사이에 결제 로직
-
-        order.updateOrderStatus(OrderStatus.ORDER_COMPLETE);
-
-        showDateUpdate(showDateId, seatIds.size());
-
-        seatHoldingService.seatHoldingUnLock(seatIds);
-
-        return order;
-    }
-
-    @Transactional
-    public List<Ticket> cancelBooking(Long showId, Long showDateId, Long userId, List<Long> ticketIds) throws InterruptedException {
-
-        ShowDate foundShowDate = showDateQueryService.getShowDate(showDateId);
-        checkCancelBookingTime(foundShowDate);
-
-        User foundUser = userQueryService.getUser(userId);
-        List<Ticket> canceledTickets = ticketCommandService.deleteTicket(foundUser, ticketIds);
-
-        showDateUpdate(showDateId, ticketIds.size());
-
-        return canceledTickets;
-
-    }
-
-    private void showDateUpdate(Long showDateId, int count) throws InterruptedException {
         String lockKey = KEY_PREFIX + showDateId;
         RLock lock = redissonClient.getFairLock(lockKey);
 
-
         if (lock.tryLock(10, TimeUnit.SECONDS)) {
             try {
-                showDateCommandService.showDateUpdate(showDateId, count);
+                Show foundShow = showQueryService.getShow(showId); // select
+                checkBookingTime(foundShow);
+
+                User foundUser = userQueryService.getUser(userId); // select
+
+                seatHoldingService.seatHoldingCheck(userId, seatIds);
+
+                ticketQueryService.checkTicketLimit(foundUser, foundShow); // select
+
+                List<Ticket> tickets = ticketCommandService.createTicket(foundUser, foundShow, seatIds); // insert
+
+                Order order = orderCommandService.createOrder(foundUser, tickets); // insert
+
+                showDateCommandService.countUpdate(showDateId, seatIds.size()); // update
+
+                seatHoldingService.seatHoldingUnLock(seatIds);
+
+                return order;
+
             } finally {
                 lock.unlock();
             }
         } else {
             throw new IllegalStateException("락 획득 실패: " + lockKey);
         }
+    }
+
+    @Transactional
+    public List<Ticket> cancelBooking(Long showId, Long showDateId, Long userId, List<Long> ticketIds) throws InterruptedException {
+
+        String lockKey = KEY_PREFIX + showDateId;
+        RLock lock = redissonClient.getFairLock(lockKey);
+
+        if (lock.tryLock(10, TimeUnit.SECONDS)) {
+            try {
+                ShowDate foundShowDate = showDateQueryService.getShowDate(showDateId);
+                checkCancelBookingTime(foundShowDate);
+
+                User foundUser = userQueryService.getUser(userId);
+                List<Ticket> canceledTickets = ticketCommandService.deleteTicket(foundUser, ticketIds);
+
+                showDateCommandService.countUpdate(showDateId, ticketIds.size());
+
+                return canceledTickets;
+
+            } finally {
+                lock.unlock();
+            }
+        } else {
+            throw new IllegalStateException("락 획득 실패: " + lockKey);
+        }
+
     }
 
     private void checkBookingTime(Show show) {
